requirements.txt
flask
faiss-cpu
sentence-transformers
requests
torch
transformers
streamlit
